+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.5] - 2020-03-15
+### Changed
+- UI changed, fixed un-matching state's name
+- UI changed, remove expand states
+- UI changed, add daily new cases, deaths cases and recovered cases for both US and CA
+
 ## [0.0.4] - 2020-03-13
 ### Changed
 - UI changed, fixed un-matching column